Session & viewComponenet

+++ TI_Net2025_DemoCleanAsp/Models/Product/ProductFilterFormDto.cs
@@ -3,8 +3,8 @@
     public class ProductFilterFormDto
     {
         public string? Name { get; set; }
-        public int? MinPrice { get; set; }
-        public int? MaxPrice { get; set; }
+        public decimal? MinPrice { get; set; }
+        public decimal? MaxPrice { get; set; }
         public int? CategoryId { get; set; }
     }
 }
